requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
onnxruntime
torch
torchvision
onnx 
